Added .gitignore file to deal with submodule/checkout.
When checking out from gh-pages to the master branch, git cannot remove the submodule "master-branch".
It raises a warning that it couldn't rmdir "master-branch".
This alone is not problematic, but the directory remains in the working directory and could erroneously
be committed. The .gitignore file makes git ignore all contents of the "master-branch" directory.